Move DailyWord component to src/components folder
