+++ src/components/daily-word/daily-word.tsx
