+++ src/pages/index.tsx
@@ -14,7 +14,7 @@ import SearchBar from "./components/input/search-bar";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { useTranslation } from "next-i18next";
 import { termFilter } from "../filter/termConfig";
-import DailyWord from "./components/daily-word/daily-word";
+import DailyWord from "../components/daily-word/daily-word";
 
 const Home: NextPage = () => {
   const { t } = useTranslation();
